Added deluge option

+++ vpnkillswitch.sh
@@ -4,14 +4,20 @@
 
 DEBUG_MODE=true
 DEBUG_PATH=/usr/vpnkillswitch/log.txt
+RESTART_DELUGE=false
 
 restartVPN()
 {
-if [ "$DEBUG_MODE" = true ] ; then echo "Stopping Deluge..." >> $DEBUG_PATH ; fi
-pkill deluged
+if [ "$RESTART_DELUGE" = true ]
+then
+	if [ "$DEBUG_MODE" = true ] ; then echo "Stopping Deluge..." >> $DEBUG_PATH ; fi
+	pkill deluged
+fi
 
 if [ "$DEBUG_MODE" = true ] ; then echo "Restarting VPN..." >> $DEBUG_PATH ; fi
 iptables -F
+#iptables -A INPUT -j ACCEPT
+#iptables -A OUTPUT -j ACCEPT
 service openvpn restart
 sleep 5
 
@@ -27,17 +33,26 @@ iptables -t mangle -F
 iptables -t mangle -X
 iptables -A INPUT -i lo -j ACCEPT
 iptables -A OUTPUT -o lo -j ACCEPT
+# Local Traffic
+iptables -A INPUT -s 127.0.0.1 -j ACCEPT
+iptables -A OUTPUT -d 127.0.0.1 -j ACCEPT
+# Broadcast
 iptables -A OUTPUT -d 255.255.255.255 -j  ACCEPT
 iptables -A INPUT -s 255.255.255.255 -j ACCEPT
-iptables -A INPUT -s 10.0.0.0/16 -d 10.0.0.0/16 -j ACCEPT
-iptables -A OUTPUT -s 10.0.0.0/16 -d 10.0.0.0/16 -j ACCEPT
+#Local network (You may need to update)
+iptables -A INPUT -s 192.168.0.0/24  -j ACCEPT
+iptables -A OUTPUT  -d 192.168.0.0/24 -j ACCEPT
+# Tunnel forwarding
 iptables -A FORWARD -i eth0 -o tun0 -j ACCEPT
 iptables -A FORWARD -i tun0 -o eth0 -j ACCEPT
 iptables -t nat -A POSTROUTING -o tun0 -j MASQUERADE
 iptables -A OUTPUT -o eth0 ! -d $WAN_IP -j DROP
 
-if [ "$DEBUG_MODE" = true ] ; then echo "Starting Deluge..." >> $DEBUG_PATH ; fi
-deluged
+if [ "$RESTART_DELUGE" = true ]
+then
+	if [ "$DEBUG_MODE" = true ] ; then echo "Starting Deluge..." >> $DEBUG_PATH ; fi
+	deluged
+fi
 }
 
 VPN=$(service openvpn status)
